float arrumado

+++ dicionario.py
@@ -8,7 +8,7 @@ def adicionarProdutoatalogo(dict, nome, quantidade, preco, descricao):
         for i in dict:
             id = int(i)
     id+=1
-    preco = int(preco)
+    preco = float(preco)
     quantidade = int(quantidade)
     
     print(f"O id é o seguinte {id}" )
@@ -29,7 +29,7 @@ def removerProduto(dict, id_delete):
         atualizarArquivo(dict)
 
 def modificarProduto(dict, id_alterar, nome, quant, valor):
-    valor = int(valor)
+    valor = float(valor)
     quant = int(quant)
     for id, produto in dict.items():
         if id == id_alterar:
